[refactor] NJT-42 Note型の切り出し
