+++ components/Providers.tsx
@@ -1,3 +1,4 @@
+// components/ui/Providers.tsx
 'use client';
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
