+++ lib/prisma.ts
@@ -1,3 +1,4 @@
+// lib/prisma.ts
 import { PrismaClient } from '@prisma/client';
 
 // PrismaClientのインスタンスをグローバルに宣言
