+++ types/index.ts
@@ -1,0 +1,3 @@
+// types/index.ts
+export * from './user';
+export * from './note';
