+++ types/note.ts
@@ -1,0 +1,7 @@
+// types/note.ts
+export type Note = {
+  id: number;
+  title: string;
+  content: string;
+  createdAt: string;
+};
