+++ types/user.ts
@@ -1,3 +1,4 @@
+// types/user.ts
 export interface User {
   userId: string;
   email: string;
